Thoughts on Instruction Prefetching
The ARM7TDMI uses instruction prefetching which allows concurrent fetch,
decode, and execution of instructions. Because of this, the PC is always
ahead of the address of the executing instruction. In our case, the
seemingly best way to handle this is to just force the PC to be ahead
before execution and guard against PC writes.